requirements-dev.txt
black==23.12.1
brotli==1.1.0
build==1.0.3
commitizen==3.13.0
ipykernel==6.29.0
mutagen==1.47.0
pip-autoremove==0.10.0
pip-chill==1.0.3
pip3-autoremove==1.2.0
pytest-cov==4.1.0
twine==4.0.2
# appnope==0.1.3 # Installed as dependency for ipykernel
# argcomplete==3.2.1 # Installed as dependency for commitizen
# asttokens==2.4.1 # Installed as dependency for stack-data
# certifi==2023.11.17 # Installed as dependency for requests
# charset-normalizer==3.3.2 # Installed as dependency for commitizen, requests
# click==8.1.7 # Installed as dependency for black
# colorama==0.4.6 # Installed as dependency for commitizen
# comm==0.2.1 # Installed as dependency for ipykernel
# coverage==7.4.0 # Installed as dependency for pytest-cov
# debugpy==1.8.0 # Installed as dependency for ipykernel
# decli==0.6.1 # Installed as dependency for commitizen
# decorator==5.1.1 # Installed as dependency for ipython
# docutils==0.20.1 # Installed as dependency for readme-renderer
# executing==2.0.1 # Installed as dependency for stack-data
# idna==3.6 # Installed as dependency for requests
# importlib-metadata==7.0.1 # Installed as dependency for commitizen, keyring, twine
# iniconfig==2.0.0 # Installed as dependency for pytest
# ipython==8.20.0 # Installed as dependency for ipykernel
# jaraco.classes==3.3.0 # Installed as dependency for keyring
# jedi==0.19.1 # Installed as dependency for ipython
# jinja2==3.1.3 # Installed as dependency for commitizen
# jupyter-client==8.6.0 # Installed as dependency for ipykernel
# jupyter-core==5.7.1 # Installed as dependency for ipykernel, jupyter-client
# keyring==24.3.0 # Installed as dependency for twine
# markdown-it-py==3.0.0 # Installed as dependency for rich
# markupsafe==2.1.3 # Installed as dependency for jinja2
# matplotlib-inline==0.1.6 # Installed as dependency for ipykernel, ipython
# mdurl==0.1.2 # Installed as dependency for markdown-it-py
# more-itertools==10.2.0 # Installed as dependency for jaraco.classes
# mypy-extensions==1.0.0 # Installed as dependency for black
# nest-asyncio==1.5.9 # Installed as dependency for ipykernel
# nh3==0.2.15 # Installed as dependency for readme-renderer
# packaging==23.2 # Installed as dependency for black, build, commitizen, ipykernel, pytest
# parso==0.8.3 # Installed as dependency for jedi
# pathspec==0.12.1 # Installed as dependency for black
# pexpect==4.9.0 # Installed as dependency for ipython
# pkginfo==1.9.6 # Installed as dependency for twine
# platformdirs==4.1.0 # Installed as dependency for black, jupyter-core
# pluggy==1.3.0 # Installed as dependency for pytest
# prompt-toolkit==3.0.43 # Installed as dependency for ipython, questionary
# psutil==5.9.7 # Installed as dependency for ipykernel
# ptyprocess==0.7.0 # Installed as dependency for pexpect
# pure-eval==0.2.2 # Installed as dependency for stack-data
# pygments==2.17.2 # Installed as dependency for ipython, readme-renderer, rich
# pyproject-hooks==1.0.0 # Installed as dependency for build
# pytest==7.4.4 # Installed as dependency for pytest-cov
# python-dateutil==2.8.2 # Installed as dependency for jupyter-client
# pyyaml==6.0.1 # Installed as dependency for commitizen
# pyzmq==25.1.2 # Installed as dependency for ipykernel, jupyter-client
# questionary==2.0.1 # Installed as dependency for commitizen
# readme-renderer==42.0 # Installed as dependency for twine
# requests==2.31.0 # Installed as dependency for requests-toolbelt, twine
# requests-toolbelt==1.0.0 # Installed as dependency for twine
# rfc3986==2.0.0 # Installed as dependency for twine
# rich==13.7.0 # Installed as dependency for twine
# six==1.16.0 # Installed as dependency for asttokens, python-dateutil
# stack-data==0.6.3 # Installed as dependency for ipython
# termcolor==2.4.0 # Installed as dependency for commitizen
# tomlkit==0.12.3 # Installed as dependency for commitizen
# tornado==6.4 # Installed as dependency for ipykernel, jupyter-client
# traitlets==5.14.1 # Installed as dependency for comm, ipykernel, ipython, jupyter-client, jupyter-core, matplotlib-inline
# urllib3==2.1.0 # Installed as dependency for requests, twine
# wcwidth==0.2.13 # Installed as dependency for prompt-toolkit
# zipp==3.17.0 # Installed as dependency for importlib-metadata
